Making Range accept double instead of integer

+++ src/Range.cpp
@@ -1,7 +1,7 @@
 #include "./Range.h"
 
 
-void Range::initializeValues(int max, int min, int default_value) {
+void Range::initializeValues(double max, double min, double default_value) {
     if(min > max) 
         throw std::invalid_argument("minimum value is larger than maximum value");
     if(default_value > max || default_value < min)
@@ -10,7 +10,7 @@ void Range::initializeValues(int max, int min, int default_value) {
     this -> min = min;
     this -> default_value = default_value; 
 }
-Range::Range(int max, int min, int default_value) {
+Range::Range(double max, double min, double default_value) {
     initializeValues(max, min, default_value);
 }
 
@@ -22,18 +22,17 @@ Range::Range(const json& rangeData) {
         throw std::invalid_argument("'min' value is missing");
     if(!rangeData.contains("default"))
         throw std::invalid_argument("'default' value is missing");
-
     initializeValues(rangeData.at("max"), rangeData.at("min"), rangeData.at("default"));
 }
-int Range::getMax() const {
+double Range::getMax() const {
     return max;
 }
 
-int Range::getMin() const {
+double Range::getMin() const {
     return min;
 }
 
-int Range::getDefault() const {
+double Range::getDefault() const {
     return default_value;
 }
